refactor: 도메인 패키지 구조 변경
* 메뉴, 테이블 도메인 각 패키지로 분리
