+++ src/main/java/domain/menu/Menu.java
@@ -2,8 +2,6 @@ package domain.menu;
 
 import java.util.Objects;
 
-import domain.Category;
-
 public class Menu {
     private static final int MIN_NUMBER = 1;
     private static final int MIN_PRICE = 0;
